feat: add a fonctionnality Maintenance

+++ src/main/java/com/gmao/gmao_project/service/EquipmentService.java
@@ -2,6 +2,8 @@ package com.gmao.gmao_project.service;
 
 import com.gmao.gmao_project.model.Equipment;
 import com.gmao.gmao_project.repository.EquipmentRepository;
+import com.gmao.gmao_project.repository.InterventionRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,9 +16,11 @@ public class EquipmentService {
     @Autowired
     private EquipmentRepository equipmentRepository;
 
+    @Autowired
+    private InterventionRepository interventionRepository;
+
     public List<Equipment> getAllEquipments() {
         List<Equipment> equipments = equipmentRepository.findAll();
-        // Ajoutez un log pour déboguer
         System.out.println("Equipments retrieved: " + (equipments != null ? equipments.size() : "null"));
         return equipments != null ? equipments : Collections.emptyList();
     }
@@ -31,6 +35,11 @@ public class EquipmentService {
     }
 
     public void deleteEquipment(Long id) {
+        if (!equipmentRepository.existsById(id)) {
+            throw new EntityNotFoundException("Équipement avec l’ID " + id + " non trouvé");
+        }
+        // Supprimer toutes les interventions liées à cet équipement
+        interventionRepository.deleteByEquipmentId(id);
         equipmentRepository.deleteById(id);
     }
 }
